refactored operators
