+++ .github/workflows/config.yml
@@ -8,7 +8,7 @@ on:
       - master
     paths:
       - airflow_logic/**
-      - dags/**
+      #- dags/**
       - plugins/**
       - .github/workflows/config.yml
   workflow_dispatch:
@@ -25,8 +25,8 @@ jobs:
           service_account_key: ${{ secrets.GCLOUD_SERVICE_KEY }}
           project_id: ${{ secrets.GCS_PROJECT }}
           export_default_credentials: true
-      - name: Copy DAGs
-        run: gsutil rsync -d -r ./dags gs://${{ secrets.GCLOUD_BUCKET_PATH }}/dags/
+      - name: Copy DAGs # TODO: verify this command
+        run: gsutil rsync -d -r ./airflow_logic/dags gs://${{ secrets.GCLOUD_BUCKET_PATH }}/dags/
       - name: Copy plugins
         run: gsutil rsync -d -r ./plugins gs://${{ secrets.GCLOUD_BUCKET_PATH }}/plugins/
       - name: Copy Airflow Logic
