+++ airflow_logic/dags/hdx_adm_dags.py
@@ -1,11 +1,10 @@
 from airflow import DAG
-
-from dags_utils.utils.config_parser import config
-from dags.dags_utils.dag_configuration import (
+from airflow_logic.dags.dags_utils.dag_configuration import (
     get_catchup,
     get_default_arguments,
     get_schedule_interval,
 )
+from dags_utils.utils.config_parser import config
 from dags_utils.utils.hdx_dags_filler import fill_hdx_adm_dag
 
 countries = config.get_countries()
