+++ airflow_logic/operators/BaseMapActionOperator.py
@@ -4,9 +4,11 @@ from typing import Callable
 from airflow.contrib.operators.kubernetes_pod_operator import KubernetesPodOperator
 from airflow.operators.python_operator import PythonOperator
 from airflow.utils.decorators import apply_defaults
+from airflow_logic.operators.operators_utils.deep_inspection import (
+    get_function_information,
+)
 
 from pipeline_plugin.pipeline_config.config_parser import config
-from pipeline_plugin.utils.deep_inspection import get_function_information
 
 
 class MapActionKubernetesPodOperator(KubernetesPodOperator):
