+++ airflow_logic/operators/HDXAdmTransformOperator.py
@@ -1,6 +1,6 @@
 from airflow.utils.decorators import apply_defaults
+from airflow_logic.operators.BaseMapActionOperator import MapActionOperator
 
-from pipeline_plugin.operators.BaseMapActionOperator import MapActionOperator
 from pipeline_plugin.transform.admn_transform import transform
